tests/check/Makefile.am: Fix CFLAGS to also pull in the gstcheck cflags from AM_CFLAGS.
Original commit message from CVS:
* tests/check/Makefile.am:
Fix CFLAGS to also pull in the gstcheck cflags from AM_CFLAGS.
* tests/check/elements/audiorate.c: (do_perfect_stream_test):
* tests/check/elements/playbin.c:
* tests/check/libs/mixer.c: (test_element_interface_supported),
(gst_implements_interface_init):
* tests/check/libs/rtp.c: (GST_START_TEST):
Fix various assignment type mismatches.

+++ tests/check/libs/rtp.c
@@ -66,8 +66,8 @@ GST_START_TEST (test_rtp_buffer)
 
   /* check ssrc offset */
   gst_rtp_buffer_set_ssrc (buf, 0xf04043C2);
-  fail_unless_equals_int (gst_rtp_buffer_get_ssrc (buf), 0xf04043c2);
-  fail_unless_equals_int (GST_READ_UINT32_BE (data + 4 + 4), 0xf04043c2);
+  fail_unless_equals_int (gst_rtp_buffer_get_ssrc (buf), (gint) 0xf04043c2);
+  fail_unless_equals_int (GST_READ_UINT32_BE (data + 4 + 4), (gint) 0xf04043c2);
   gst_rtp_buffer_set_ssrc (buf, 0);
   fail_unless_equals_int (gst_rtp_buffer_get_ssrc (buf), 0);
   fail_unless_equals_int (GST_READ_UINT32_BE (data + 4 + 4), 0);
